RequirementMachine: Split off Rule.cpp from RewriteSystem.cpp
