+++ lib/AST/RequirementMachine/RewriteSystem.h
@@ -16,11 +16,11 @@
 #include "swift/AST/Requirement.h"
 #include "swift/AST/TypeCheckRequests.h"
 #include "llvm/ADT/DenseSet.h"
-#include "llvm/ADT/PointerUnion.h"
 
 #include "Debug.h"
 #include "Diagnostics.h"
 #include "RewriteLoop.h"
+#include "Rule.h"
 #include "Symbol.h"
 #include "Term.h"
 #include "Trie.h"
@@ -37,185 +37,6 @@ namespace rewriting {
 class PropertyMap;
 class RewriteContext;
 class RewriteSystem;
-
-/// A rewrite rule that replaces occurrences of LHS with RHS.
-///
-/// LHS must be greater than RHS in the linear order over terms.
-///
-/// Out-of-line methods are documented in RewriteSystem.cpp.
-class Rule final {
-  Term LHS;
-  Term RHS;
-
-  /// The written requirement ID, which can be used to index into the
-  /// \c WrittenRequirements array in the rewrite system to retrieve
-  /// the structural requirement.
-  Optional<unsigned> requirementID;
-
-  /// A 'permanent' rule cannot be deleted by homotopy reduction. These
-  /// do not correspond to generic requirements and are re-added when the
-  /// rewrite system is built.
-  unsigned Permanent : 1;
-
-  /// An 'explicit' rule is a generic requirement written by the user.
-  unsigned Explicit : 1;
-
-  /// An 'LHS simplified' rule's left hand side was reduced via another rule.
-  /// Set by simplifyLeftHandSides().
-  unsigned LHSSimplified : 1;
-
-  /// An 'RHS simplified' rule's right hand side can be reduced via another rule.
-  /// Set by simplifyRightHandSides().
-  unsigned RHSSimplified : 1;
-
-  /// A 'substitution simplified' rule's left hand side contains substitutions
-  /// which can be reduced via another rule.
-  /// Set by simplifyLeftHandSideSubstitutions().
-  unsigned SubstitutionSimplified : 1;
-
-  /// A 'redundant' rule was eliminated by homotopy reduction. Redundant rules
-  /// still participate in term rewriting, but they are not part of the minimal
-  /// set of requirements in a generic signature.
-  unsigned Redundant : 1;
-
-  /// A 'conflicting' rule is a property rule which cannot be satisfied by any
-  /// concrete type because it is mutually exclusive with some other rule.
-  /// An example would be a pair of concrete type rules:
-  ///
-  ///    T.[concrete: Int] => T
-  ///    T.[concrete: String] => T
-  ///
-  /// Conflicting rules are detected in property map construction, and are
-  /// dropped from the minimal set of requirements.
-  unsigned Conflicting : 1;
-
-public:
-  Rule(Term lhs, Term rhs)
-      : LHS(lhs), RHS(rhs) {
-    Permanent = false;
-    Explicit = false;
-    LHSSimplified = false;
-    RHSSimplified = false;
-    SubstitutionSimplified = false;
-    Redundant = false;
-    Conflicting = false;
-  }
-
-  const Term &getLHS() const { return LHS; }
-  const Term &getRHS() const { return RHS; }
-
-  Optional<unsigned> getRequirementID() const {
-    return requirementID;
-  }
-
-  void setRequirementID(Optional<unsigned> requirementID) {
-    assert(!getRequirementID().hasValue());
-    this->requirementID = requirementID;
-  }
-
-  Optional<Symbol> isPropertyRule() const;
-
-  const ProtocolDecl *isProtocolConformanceRule() const;
-
-  const ProtocolDecl *isAnyConformanceRule() const;
-
-  bool isIdentityConformanceRule() const;
-
-  bool isProtocolRefinementRule() const;
-
-  bool isCircularConformanceRule() const;
-
-  /// See above for an explanation of these predicates.
-  bool isPermanent() const {
-    return Permanent;
-  }
-
-  bool isExplicit() const {
-    return Explicit;
-  }
-
-  bool isLHSSimplified() const {
-    return LHSSimplified;
-  }
-
-  bool isRHSSimplified() const {
-    return RHSSimplified;
-  }
-
-  bool isSubstitutionSimplified() const {
-    return SubstitutionSimplified;
-  }
-
-  bool isRedundant() const {
-    return Redundant;
-  }
-
-  bool isConflicting() const {
-    return Conflicting;
-  }
-
-  bool containsUnresolvedSymbols() const {
-    return (LHS.containsUnresolvedSymbols() ||
-            RHS.containsUnresolvedSymbols());
-  }
-
-  Optional<Identifier> isProtocolTypeAliasRule() const;
-
-  bool isDerivedFromConcreteProtocolTypeAliasRule() const;
-
-  void markLHSSimplified() {
-    assert(!LHSSimplified);
-    LHSSimplified = true;
-  }
-
-  void markRHSSimplified() {
-    assert(!RHSSimplified);
-    RHSSimplified = true;
-  }
-
-  void markSubstitutionSimplified() {
-    assert(!SubstitutionSimplified);
-    SubstitutionSimplified = true;
-  }
-
-  void markPermanent() {
-    assert(!Explicit && !Permanent &&
-           "Permanent and explicit are mutually exclusive");
-    Permanent = true;
-  }
-
-  void markExplicit() {
-    assert(!Explicit && !Permanent &&
-           "Permanent and explicit are mutually exclusive");
-    Explicit = true;
-  }
-
-  void markRedundant() {
-    assert(!Redundant);
-    Redundant = true;
-  }
-
-  void markConflicting() {
-    // It's okay to mark a rule as conflicting multiple times, but it must not
-    // be a permanent rule.
-    assert(!Permanent && "Permanent rule should not conflict with anything");
-    Conflicting = true;
-  }
-
-  unsigned getDepth() const;
-
-  unsigned getNesting() const;
-
-  Optional<int> compare(const Rule &other, RewriteContext &ctx) const;
-
-  void dump(llvm::raw_ostream &out) const;
-
-  friend llvm::raw_ostream &operator<<(llvm::raw_ostream &out,
-                                       const Rule &rule) {
-    rule.dump(out);
-    return out;
-  }
-};
 
 /// Result type for RequirementMachine::computeCompletion().
 enum class CompletionResult {
